requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
autocorrect
webcolors
numpy
pandas
seaborn
pyplot-themes
cv2
urllib3_1_26_2 
picklable-itertools 
random2
10S
glob2
imutils
sklearn_utils
pycopy-urllib.request
pycopy-warnings
tensorflow==2.11
--upgrade Pillow
python-csv
micropython-fileinput
python-math
Flask